% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_graph}
\alias{create_graph}
\title{Create empty graph}
\usage{
create_graph()
}
\value{
An object of type \code{chickenwire} which is a graph. In this case an empty
graph to which edges and vertices can be added using \code{\link{add_edges}} 
and \code{\link{add_vertices}}.
}
\description{
Create empty graph
}
