% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_graph_old}
\alias{create_graph_old}
\title{Create graph}
\usage{
create_graph_old(vertex_id, edge_src, edge_dst, edge_weight = 1,
  edge_type = 1L)
}
\arguments{
\item{vertex_id}{vector with vertex id's.}

\item{edge_src}{vector with id's of the starting nodes of the edges. Should
be of the same type as \code{vertex_id}.}

\item{edge_dst}{vector with id's of the end nodes of the edges. Should
be of the same type as \code{vertex_id}.}

\item{edge_weight}{numeric vector with edge weights (optional).}

\item{edge_type}{vector with edge types (optional).}
}
\description{
Create graph
}
\details{
It is possible to call \code{create_graph} without any argument. An empty
graph is then created to which edges and vertices can be added using 
\code{\link{add_edges}} and \code{\link{add_vertices}}. 

Vertex id's are stored internally as integers. 
When \code{vertex_id} is of type \code{character} or \code{factor} it is 
converted to \code{integer}. \code{edge_src} and \code{edge_dst} are also
recoded to integer. This is important when later adges or vertices are added
using \code{\link{add_edges}} or \code{\link{add_vertices}}. In that case
it is necessary to do the conversion to integer manually before calling 
\code{create_graph}.
}
