% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_average.R
\name{local_average}
\alias{local_average}
\title{Random Walk
Local network averages}
\usage{
local_average(graph_id, vertex_id, vertex_value, vertex_weight = 1)
}
\description{
Random Walk
Local network averages
}
